+++ javascript/packages/eslint-config-chotot-base/rules/node.js
@@ -4,40 +4,18 @@ module.exports = {
   },
 
   rules: {
-    // enforce return after a callback
-    'callback-return': 'off',
-
-    // require all requires be top-level
-    // https://eslint.org/docs/rules/global-require
-    'global-require': 'error',
-
-    // enforces error handling in callbacks (node environment)
-    'handle-callback-err': 'off',
-
     // disallow use of the Buffer() constructor
     // https://eslint.org/docs/rules/no-buffer-constructor
+    // deprecated: moved to eslint-plugin-node
     'no-buffer-constructor': 'error',
 
-    // disallow mixing regular variable and require declarations
-    'no-mixed-requires': ['off', false],
-
     // disallow use of new operator with the require function
+    // deprecated: moved to eslint-plugin-node
     'no-new-require': 'error',
 
     // disallow string concatenation with __dirname and __filename
     // https://eslint.org/docs/rules/no-path-concat
+    // deprecated: moved to eslint-plugin-node
     'no-path-concat': 'error',
-
-    // disallow use of process.env
-    'no-process-env': 'off',
-
-    // disallow process.exit()
-    'no-process-exit': 'off',
-
-    // restrict usage of specified node modules
-    'no-restricted-modules': 'off',
-
-    // disallow use of synchronous methods (off by default)
-    'no-sync': 'off',
   }
 };