+++ javascript/packages/eslint-config-chotot/rules/react.js
@@ -65,7 +65,10 @@ module.exports = {
 
     // Enforce a specific function type for function components
     // https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/function-component-definition.md
-    'react/function-component-definition': 'off',
+    'react/function-component-definition': ['warn', {
+      namedComponents: 'arrow-function',
+      unnamedComponents: 'arrow-function',
+    }],
 
     // Enforce boolean attributes notation in JSX
     // https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/jsx-boolean-value.md
@@ -399,11 +402,6 @@ module.exports = {
       afterOpening: 'never',
       beforeClosing: 'never',
     }],
-
-    // Enforce spaces before the closing bracket of self-closing JSX elements
-    // https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/jsx-space-before-closing.md
-    // Deprecated in favor of jsx-tag-spacing
-    'react/jsx-space-before-closing': ['off', 'always'],
 
     // Prevent usage of Array index in keys
     // https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/no-array-index-key.md